project: changed priority to use shorts instead of ints

+++ api/src/main/java/wtf/mizu/kawa/api/Priority.java
@@ -8,36 +8,36 @@ package wtf.mizu.kawa.api;
  */
 public class Priority {
     /**
-     * The lowest priority, will be consumed after everyone else.
+     * The lowest priority, will usually be consumed last.
      */
-    public static final int LOWEST = -1000;
+    public static final short LOWEST = -1000;
 
     /**
      * A low priority, will be consumed a bit later than most other
      * subscriptions.
      */
-    public static final int LOW = -500;
+    public static final short LOW = -500;
 
     /**
      * The default priority.
      */
-    public static final int DEFAULT = 0;
+    public static final short DEFAULT = 0;
 
     /**
      * A high priority, will be consumed earlier than most other subscriptions.
      */
-    public static final int HIGH = 500;
+    public static final short HIGH = 500;
 
     /**
      * The highest priority, will ensure its subscription to be consumed first.
      */
-    public static final int HIGHEST = 1000;
+    public static final short HIGHEST = 1000;
 
     /**
      * An array containing the possible default values,
      * ordered from lowest to highest.
      */
-    public static final int[] values = {
+    public static final short[] values = {
             LOWEST,
             LOW,
             DEFAULT,
